+++ examples/distributed_simulation/world_manager.py
@@ -2,15 +2,21 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+import sys
+
+from mango import addr
+
 from assume import MarketConfig, World
 from assume.common.forecasts import NaiveForecast
 
 from .config import (
+    agent_addresses,
     db_uri,
     index,
     manager_protocol_addr,
     market_operator_addr,
     marketdesign,
+    use_mqtt,
     worker,
 )
 
@@ -25,8 +31,6 @@ async def create_worker(
     world.add_market_operator(id=market_operator_addr.aid)
     for market_config in marketdesign:
         world.add_market(market_operator_addr.aid, market_config)
-
-    world.add_unit_operator(f"my_operator{i}")
 
     world.add_unit_operator(f"my_demand{i}")
     world.add_unit(
@@ -48,4 +52,25 @@ if __name__ == "__main__":
     world = World(
         database_uri=db_uri, addr=manager_protocol_addr, distributed_role=True
     )
-    world.loop.run_until_complete(worker(world, marketdesign, create_worker))
+
+    # command line option can set the agent_addresses of the distributed simulation accordingly
+    # one can use this script as follows for TCP:
+    # python3 -m examples.distributed_simulation.world_manager 9098 9099
+    # and for MQTT:
+    # python3 -m examples.distributed_simulation.world_manager agent agent2
+    if len(sys.argv) > 1:
+        agent_addresses = []
+        for address in sys.argv[1:]:
+            if not use_mqtt:
+                host, port = address.split(":")
+                agent_address = (host, int(port))
+            else:
+                agent_address = address
+            agent_addresses.append(addr(agent_address, "clock_agent"))
+        print("new agent addresses are", agent_addresses)
+    try:
+        if world.distributed_role:
+            world.addresses.extend(agent_addresses)
+        world.loop.run_until_complete(worker(world, marketdesign, create_worker))
+    except Exception as e:
+        print(e)